+++ src/main/java/com/bokmcdok/butterflies/world/item/ButterflyNetItem.java
@@ -53,6 +53,8 @@ public class ButterflyNetItem extends Item implements ButterflyContainerItem {
     //  TODO: Remove this item
     public static final String FULL_NAME = "butterfly_net_full";
 
+    private static final String NAME = "item.butterflies.butterfly_net";
+
     //  The index of the butterfly species.
     private final int butterflyIndex;
 
@@ -111,6 +113,18 @@ public class ButterflyNetItem extends Item implements ButterflyContainerItem {
     @Override
     public ItemStack getCraftingRemainingItem(ItemStack itemStack) {
         return new ItemStack(ItemRegistry.BUTTERFLY_NET.get());
+    }
+
+    /**
+     * Overridden so we can use a single localisation string for all instances.
+     * @param itemStack The stack to get the name for.
+     * @return The description ID, which is a reference to the localisation
+     *         string.
+     */
+    @NotNull
+    @Override
+    public Component getName(@NotNull ItemStack itemStack) {
+        return Component.translatable(NAME);
     }
 
     /**